Move DgraphBundledHelper to test_dgraph_bulk.py, and add more exception logs
